% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell.props}
\alias{cell.props}
\title{cell.props}
\format{A matrix of 150 samples X 22 cell types.}
\usage{
cell.props
}
\description{
Estimated cell proportions. Estimated using LM22 cell signatures
}
\keyword{datasets}

