% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell.sigs}
\alias{cell.sigs}
\title{cell.sigs}
\format{A matrix of 22 cell types x 490 genes, log2 scale data}
\source{
\url{https://cibersort.stanford.edu/}
}
\usage{
cell.sigs
}
\description{
A subset of the LM22 cell signatures
}
\keyword{datasets}

