% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.resids.synthetic.R
\name{custom.resids.synthetic}
\alias{custom.resids.synthetic}
\title{custom.resids.synthetic}
\usage{
custom.resids.synthetic(mse2model.vec, groups, diff.2.model.vec, base.expr.vec,
  adjuster = 1, n.cells)
}
\arguments{
\item{mse2model.vec}{The vector of mean squared errors (MSEs) to model}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{diff.2.model.vec}{Vector of cell type-specific differential expressions to create.}

\item{base.expr.vec}{Vector of base (control group) cell type specific expressions to model}

\item{adjuster}{A scaling factor used on the resulting standard deviation of residuals.}
}
\value{
syn.resids  Synthetic normal residuals based on standard deviations of control and cases residuals from resids.gene.j.
}
\description{
Called from user lrcde wrapper script (see example 'lrcde.wrapper.sims.R').
}
\examples{
custom.resids.synthetic( mse2model.vec, groups, diff.2.model.vec, base.expr.vec, adjuster=1  )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

