% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.sim.cell.expr.R
\name{custom.sim.cell.expr}
\alias{custom.sim.cell.expr}
\title{custom.sim.cell.expr}
\usage{
custom.sim.cell.expr(num.cells = 5, base.expr.vec, diff.2.model.vec, cell.p,
  len.mse.vec)
}
\arguments{
\item{num.cells}{Number of cell types to simulate.}

\item{base.expr.vec}{Control group expression level to simulate at the cell type-specific level.}

\item{cell.p}{Number of cell type to alter between controls and cases for simulation purposes.}

\item{diff.expr.vec}{Absolute difference between control group and cases group to simulate at the cell type-specific level.}
}
\value{
cell.expr Cell expressions matrix.
}
\description{
Called from a user wrapper script (see example 'lrcde.wrapper.sims.R').
}
\examples{
custom.sim.cell.expr( num.cells.2.simulate, base.expression.vector, difference.2.simulate.vector, cell.2.modify )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

