% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.cell.props.R
\name{decon.cell.props}
\alias{decon.cell.props}
\title{decon.cell.props}
\usage{
decon.cell.props(het.sub, cell.sigs)
}
\arguments{
\item{het.sub}{Heterogeneous gene expression matrix, _n_ samples by _m_ genes 
(rows by columns). Required. Note that only a subset of heterogeneous measres
is used, with the same number and order of genes as in the cell signature matrix}

\item{cell.sigs}{Cell-specific gene expression signatures, _p_ cell types by 
_m_ genes (rows by columns). Required. Note that both heterogeneous matrix and 
cell signature matrix should be on the same scale, log2 or decimal}
}
\value{
Estimated cell proportions, _n_ samples by _m_ cell types (rows by columns)
}
\description{
Call this function to estimate cell proportions given hetergenous expressions and cell signatures.
}
\details{
decon.cell.props is an auxillary function meant to be called directly by user in order to estimate cell proportions given cell gene signatures.
This assumes that G and cell.sigs have been previously made conformal with identical gene names/labels.
}
\examples{

\dontrun{
# Loads the heterogeneous matrix (subset) and cell signatures
data("lrcde.demo") # Included in the LRCDE function
# Estimate cell proportions. 
cell.props <- decon.cell.props(het.mtx, cell.sigs)
}
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

