% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.permuted.p.values.R
\name{get.permuted.p.values}
\alias{get.permuted.p.values}
\title{get.permuted.p.values}
\usage{
get.permuted.p.values(diff.perms.func, diffs.func, alt.is, cell.p)
}
\arguments{
\item{diff.perms.func}{Differential expression estimates from permutations (a 3D array: permutations by cell types by genomic sites).}

\item{diffs.func}{Differential expression estimates from group label non-permuted observations.}

\item{alt.is}{Defaults to 'greater' to test for up-regulation of a genomic site.  All differences are modled by absolute value, so default is 'greater'.}

\item{cell.p}{Integer of cell type being tested for differences between truth (diffs.func) and permuted differences (diff.perms.func).}
}
\value{
sig.perms.func  Vector of permutation p.values for each genomic site.
}
\description{
Called from sim_driver_script
}
\examples{
get.permuted.p.values(  diff.perms.func, diffs.func, alt.is, cell.p  )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

