% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.resids.synthetic.R
\name{get.resids.synthetic}
\alias{get.resids.synthetic}
\title{get.resids.synthetic}
\usage{
get.resids.synthetic(resids.gene.j, groups, num.genes.desired)
}
\arguments{
\item{resids.gene.j}{The residuals for both study groups for the genomic site j of interest.}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{num.genes.desires}{Number of genomic sites to simulate based on genomic site j.}
}
\value{
syn.resids  Synthetic normal residuals based on standard deviations of control and cases residuals from resids.gene.j.
}
\description{
Called from sim_driver_script.  Not called directly.
}
\examples{
lrcde(het.obs, cell.props, groups, n.perms=1000, file.ext="my.results", cell.vector.to.analyze, proj.root="Project.directory", output.dir="output.dir" )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

