% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.wise.sample.size.R
\name{group.wise.sample.size}
\alias{group.wise.sample.size}
\title{group.wise.sample.size}
\usage{
group.wise.sample.size(groups)
}
\arguments{
\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}
}
\value{
n  A vector containing control and cases sample sizes.
}
\description{
Utility function to keep from having to pass group-wise sample sizes between functions.
}
\examples{
group.wise.sample.size( groups )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

