% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/het.from.synthetic.R
\name{het.from.synthetic}
\alias{het.from.synthetic}
\title{het.from.synthetic}
\usage{
het.from.synthetic(cell.props, cell.expr, resids, groups)
}
\arguments{
\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{cell.expr}{Matrix of simulated cell type-specific expressions (returned from sim.cell.expr()).}

\item{resids}{The simulated residuals returned from get.resids.synthetic().}

\item{groups}{Vector of 1's and 2's indicating group membership (controls and cases).}
}
\value{
het.obs The simulated samples by genomic sites heterogeneous observations matrix with simulated residuals applied.
}
\description{
Called from sim_driver_script.
}
\examples{
het.from.synthetic( cell.props, simulated.cell.expression, simulated.residuals, groups )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

