% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{het.mtx}
\alias{het.mtx}
\title{het.mtx}
\format{A matrix of 150 samples x 490 genes, log2 scale data}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE65391}
}
\usage{
het.mtx
}
\description{
A subset of heterogeneous gene expression measures
}
\keyword{datasets}

