% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{het.pheno}
\alias{het.pheno}
\title{het.pheno}
\format{A factor of length 150, 72 healthy and 78 SLE subjects}
\usage{
het.pheno
}
\description{
An indicator vector of healthy/disease state
}
\keyword{datasets}

