% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LM22}
\alias{LM22}
\title{LM22}
\format{A data frame of 547 genes x gene symbol + 22 cell types, decimal scale data}
\source{
\url{https://cibersort.stanford.edu/}
}
\usage{
LM22
}
\description{
The LM22 cell signatures
}
\keyword{datasets}

