% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.cell.expr.R
\name{sim.cell.expr}
\alias{sim.cell.expr}
\title{sim.cell.expr}
\usage{
sim.cell.expr(num.cells, num.genes = 5000, base.expr = 5, diff.expr = 1,
  prop.2.fold = 0.5, cell.p)
}
\arguments{
\item{num.cells}{Number of cell types to simulate.}

\item{num.genes}{Number of genomic sites to simulate.}

\item{base.expr}{Control group expression level to simulate at the cell type-specific level.}

\item{diff.expr}{Absolute difference between control group and cases group to simulate at the cell type-specific level.}

\item{prop.2.fold}{Proportion of simulate genomic sites to "fold" or alter between controls and cases.}

\item{cell.p}{Number of cell type to alter between controls and cases for simulation purposes.}
}
\value{
List containing group-wise matrix of simulated cell type-specific expressions (rbinded as controls and cases.) and "truth" vector of 0's and 1's indicating which genomic sites have been altered for cell type p.
}
\description{
Called from sim_driver_script.
}
\examples{
sim.cell.expr(num.cells.2.simulate, num.genes.2.simulate, base.expression, difference.2.simulate, proportion.of.sim.genes.2.fold, cell.2.modify )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

