% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_driver_script.R
\name{sim_driver_script}
\alias{sim_driver_script}
\title{sim_driver_script}
\usage{
sim_driver_script(decon.list, one.gene, cell.props, groups, medCntr = TRUE,
  stdz = FALSE, nonNeg = TRUE, GET.FOLDS, n.perms = 1000,
  n.genes = 1000, setSeed = FALSE, seed2set = 11221963, output.file,
  cells.to.analyze, gene.no, FEEDBACK = TRUE, repeats = 1)
}
\arguments{
\item{decon.list}{Output list returned by do.dual.decon().  Contains diff.expr estimates, residuals from lm fits, and "lm" fit objects.}

\item{one.gene}{Single genomic site across all samples from original input heterogeneous observations matrix.}

\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{groups}{Vector of 1's and 2's indicating group membership (controls and cases).}

\item{GET.FOLDS}{Boolean idicating whether to take the absolute difference between group-wise cell type-specific estimate or to take ratio between cases and controls (for fold change estimate).}

\item{n.perms}{Integer indicating number of group label permutations to perform during genomic site simulations.  Higher numbers result in increased run time.}

\item{n.genes}{Integer indicating number of genomic sites to simulate.  Higher numbers result in increased run time.}

\item{setSeed}{Boolean indicating whether or not to set the R random seed.  Defaults to TRUE across all simulations for replicable results.}

\item{seed2set}{If setSeed=TRUE then use this seed.  Defaults to 11221963.}

\item{output.file}{Name of .csv file containing power analysis results data frame.}

\item{cells.to.analyze}{A vector of integers indicating whic of the cell types in cell.props to analyze.  Defaults to running all cell types if left empty.  Passed in from lrcde function.}

\item{gene.no}{Integer of genomic site being analyzed (row number of input heterogenous data set being analyzed).  Used for output of progress indicator.}

\item{FEEDBACK}{Boolean indicating whether to output progess indication to console.  Default is TRUE.}

\item{repeats}{Integer indicating number of times to repeat the simulation per gene.  Meant to allow for empirical measurement of AUC and p.value variability so that confidence intervals can be estimated for these two outcome measures.}
}
\value{
Nothing.  Outputs .csv file containing results of power analysis.  Ou.
}
\description{
Call this function to run entire functionality of LRCDE.  Recommend using LRCDE wrapper function to call this.
}
\details{
sim_driver_script is not meant to be called directly by user.  It is meant to be called from the lrcde script.
This script takes as input a single genomic site at a time, 'one.gene', and takes regression parameters from the results fed in by lrcde.
The regression results for 'one.gene' are then used to perform a simulation run with n.perms permutations of group labels.
This simulation is performed for 'one.gene' for each cell indicated in 'cells.to.analyze'.
The power analysis returned by the simulation is then appended as a row to the output table in output.file.csv.
}
\examples{
lrcde(het.obs, cell.props, groups, n.perms=1000, file.ext="my.results", cell.vector.to.analyze, proj.root="Project.directory", output.dir="output.dir" )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

