% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic.cell.proportions.R
\name{synthetic.cell.proportions}
\alias{synthetic.cell.proportions}
\title{synthetic.cell.proportions}
\usage{
synthetic.cell.proportions(props.2.model, groups, cell.p, setSeed = FALSE,
  seed2set = 112219623)
}
\arguments{
\item{props.2.model}{Original cell proportions supplied to model.}

\item{groups}{Vector of 1's and 2's indicating group membership.}

\item{cell.p}{Cell of interest.  This one will have the targeted SD (target.sd).}

\item{setSeed}{Boolean indicating whether to set the R random number seed.}

\item{seed2set}{Integer to used for random seed.  Defaults to 11221963.}
}
\value{
cell.props A samples by cell types matrix of simulated relative cell proportions.
}
\description{
Creates synthetic cell proportions which targets both a specific cell proportion standard deviation for cell 1 (the cell used for simulation purposes) and condition number over entire cell proportion matrix.
}
\details{
NOTE: DO not attempt to specify a condition number less than 15.  Function will crash.
}
\examples{
synthetic.cell.proportions( start.prop, n.cells, n, target.sd, target.cd=50  )

}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

