% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.props.target.R
\name{cell.props.target}
\alias{cell.props.target}
\title{cell.props.target}
\usage{
cell.props.target(n.cells, n.samps, target.sd, target.cd)
}
\arguments{
\item{n.samps}{Sample size for cell proportions matrix to create.}

\item{target.sd}{Standard deviation to create across samples for cell of interest.}

\item{target.cd}{Condition number for whole cell proportions matrix.}
}
\value{
cell.props An n.samps by cell types matrix of simulated relative cell proportions.
}
\description{
Creates synthetic cell proportions which targets both a specific cell proportion standard deviation for cell 1 (the cell used for simulation purposes) and condition number over entire cell proportion matrix.
}
\details{
NOTE: DO not attempt to specify a condition number less than 15.  Function will crash.
}
\examples{
cell.props.target(  n.samps, target.sd, target.cd )

}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

