% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom.resids.synthetic.R
\name{custom.resids.synthetic}
\alias{custom.resids.synthetic}
\title{custom.resids.synthetic}
\usage{
custom.resids.synthetic(mse2model.vec, groups, diff.2.model.vec, base.expr.vec,
  adjuster = 1)
}
\arguments{
\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{resids.gene.j}{The residuals for both study groups for the genomic site j of interest.}

\item{num.genes.desires}{Number of genomic sites to simulate based on genomic site j.}
}
\value{
syn.resids  Synthetic normal residuals based on standard deviations of control and cases residuals from resids.gene.j.
}
\description{
Called from user lrcde wrapper script (see example 'lrcde.wrapper.sims.R').
}
\examples{
custom.resids.synthetic(het.obs, cell.props, groups, n.perms=1000, file.ext="my.results", cell.vector.to.analyze, proj.root="Project.directory", output.dir="output.dir" )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

