% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decon.cell.props.R
\name{decon.cell.props}
\alias{decon.cell.props}
\title{decon.cell.props}
\usage{
decon.cell.props(het.sub, cell.sigs, LOGGED)
}
\arguments{
\item{het.sub}{Should be samples by genomic site (rows by columns).  The samples by genomic measures heterogeneous observations matrix.}

\item{cell.sigs}{Should be cell types by genes (rows by columns).}

\item{LOGGED}{A boolean indicating whether the heterogeneous observations have been log transformed.}
}
\value{
cell.props The deconvolved (estimated cell proportions)
}
\description{
Call this function to estimate cell proportions given hetergenous expressions and cell signatures.
}
\details{
decon.cell.props is an auxillary function meant to be called directly by user in order to estimate cell proportions given generic cell gene signatures.
This assumes that G and cell.sigs have been previously made conformal with identical gene names/labels.
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

