% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.decon.power.R
\name{do.decon.power}
\alias{do.decon.power}
\title{do.decon.power}
\usage{
do.decon.power(decon.list, groups)
}
\arguments{
\item{decon.list}{List output by 'do.dual.decon' (or 'do.single.decon') function.  Containst lm fit object with regression results per group.}
}
\value{
power.list  Results of power analysis.
}
\description{
This function is called by the lrcde function.  Not meant to be called directly by user.
}
\details{
This function is the main analysis part of the LRCDE package.  Here is where critical difference between control and case groups is estimated.
}
\examples{
do.decon.power( decon.list )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/ERGmisc}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

