% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.wise.sample.size.R
\name{group.wise.sample.size}
\alias{group.wise.sample.size}
\title{group.wise.sample.size}
\usage{
group.wise.sample.size(groups)
}
\arguments{
\item{groups}{A vector of identifiers indicating group membership per sample.
Should have the same length as the number of rows in het.sub and cell.props.}
}
\value{
n  A vector containing group-wise sample sizes.
}
\description{
Utility function to avoid passing group-wise sample sizes between functions.
}
\examples{
\dontrun{
group.wise.sample.size(group = rep(1:3, each = 5)) # Three groups, 5 samples each
}
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

