% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcde.permutations.R
\name{lrcde.permutations}
\alias{lrcde.permutations}
\title{lrcde.permutations}
\usage{
lrcde.permutations(het.sub, cell.props, groups, medCntr = FALSE,
  stdz = FALSE, nonNeg = TRUE, method = "dual", direction = "two.sided",
  n.perms = 100, cell.p = 1, truth)
}
\arguments{
\item{het.sub}{Should be samples by genomic site (rows by columns).  The samples by genomic measures heterogeneous observations matrix.}

\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{medCntr}{Boolean indicating whether to mean center differential expression estimates.}

\item{stdz}{Boolean indicatin whether to scale differential expression estimates with their pooled adjusted standard deviation}

\item{nonNeg}{Boolean indicating whether to force cell type-specific estimates to be non-negative (TRUE) or not (FALSE).}

\item{method}{Only "dual" is implemented in this version. This should be one of "dual" (csSAM method), or "ridge".  Default is "dual".  Specifies which type of regression deconvolution to perform.}

\item{direction}{Should be one of "two.sided", "up", or "down".  Which direction to test for cell type-specific expression changes.}

\item{n.perms}{Number of permutations of group labels to perform}

\item{cell.p}{The simulation differentiated target cell.}

\item{truth}{The truth indicator vector of which simulated "genes" in the target cell are differentiated between cases and controls.}
}
\value{
List containing data.frame (total.frame) of analysis results and a list of parameter values supplied to lrcde function (arg.used).
}
\description{
Call this function to run entire functionality of LRCDE.
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}

