% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcde.R
\name{lrcde}
\alias{lrcde}
\title{lrcde}
\usage{
lrcde(het.sub, cell.props, groups, output.file = "LRCDE_power_analysis.csv",
  VERBOSE = TRUE, medCntr = FALSE, stdz = FALSE, nonNeg = TRUE,
  method = "dual", direction = "two.sided")
}
\arguments{
\item{het.sub}{Should be samples by genomic site (rows by columns).  The samples by genomic measures heterogeneous observations matrix.}

\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{output.file}{file name to output the results. Default: LRCDE_power.analysis.csv.}

\item{VERBOSE}{Boolean indicating whether to output progess indication to console.  Default is TRUE.}

\item{medCntr}{Boolean indicating whether to mean center differential expression estimates.}

\item{stdz}{Boolean indicatin whether to scale differential expression estimates with their pooled adjusted standard deviation}

\item{nonNeg}{Boolean indicating whether to force cell type-specific estimates to be non-negative (TRUE) or not (FALSE).}

\item{method}{Only "dual" is implemented in this version. This should be one of "dual" (csSAM method), or "ridge".  Default is "dual".  Specifies which type of regression deconvolution to perform.}

\item{direction}{Should be one of "two.sided", "up", or "down".  Which direction to test for cell type-specific expression changes.}
}
\value{
List containing data.frame (total.frame) of analysis results and a list of parameter values supplied to lrcde function (arg.used).
}
\description{
Call this function to run entire functionality of LRCDE.
}
\details{
The lrcde function is meant to be called directly by user.  It is the entry point for the cell type-specific differential expression power analysis.
This is mainly a wrapper script for the deconvolution step and subsequent call to power analysis function (do.decon.power).
Default behavior is to write output to LRCDE_power.analysis.csv in the current working directory.
}
\examples{
\dontrun{
library("lrcde")
library(CellMix)
library(GEOquery)

gedDataInfo() # Data sets in CellMix to select from
mix <- ExpressionMix("GSE19830") # Example dataset with gene expression form mixture of tissues
p.sub <- subset(pData(mix), select = c(Liver, Brain, Lung, characteristics_ch1))

het <- t(exprs(mix))  # Heterogeneous expression matrix
cell.props <- t(coef(mix)) # Cell proportions matrix

apply(cell.props, 2, sd) # SD of cell proportions should not be 0
# Cell proportion standard deviation > 0.06 across samples should
# produce reasonable differential detection power.  
apply(cell.props, 1, sum)  # These sum perfectly to 1.

# Create (artificially) two groups, Brain and Lung vs. Liver
p.sub.1 <- p.sub[(p.sub$Brain + p.sub$Lung) > (p.sub$Liver), ]
# Heterogeneous matrixes for each group
het.1 <- het[rownames(het) \%in\% rownames(p.sub.1), ]
het.2 <- het[!(rownames(het) \%in\% rownames(p.sub.1)), ]
het2use <- rbind(het.1, het.2) # Combined matrix
# Cell proportions for each group
cells.1 <- cell.props[rownames(cell.props) \%in\% rownames(p.sub.1), ]
cells.2 <- cell.props[!(rownames(cell.props) \%in\% rownames(p.sub.1)), ]
cell.props <- rbind(cells.1, cells.2) # Combined cell proportions
# Group vector
groups <- c(rep(1, dim(het.1)[1]), rep(2, dim(het.2)[1]))

# Now apply the lrcde function in order to deconvolve cell
# type-specific expression and perform power analysis on differential
# expression detection:

power.analysis.df <- lrcde(het2use, 
                           cell.props, 
                           groups, 
                           output.file = "LRCDE_power_analysis_results.csv", 
                           VERBOSE = TRUE, 
                           medCntr = FALSE, 
                           stdz = FALSE, 
                           nonNeg = TRUE, 
                           method = "dual", 
                           direction = "two.sided")
}
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

