% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcde.R
\name{lrcde}
\alias{lrcde}
\title{lrcde}
\usage{
lrcde(het.sub, cell.props, groups, output.file = "LRCDE_power_analysis",
  FEEDBACK = TRUE, medCntr = FALSE, stdz = FALSE, nonNeg = TRUE,
  method = "dual", direction = "two.sided")
}
\arguments{
\item{het.sub}{Should be samples by genomic site (rows by columns).  The samples by genomic measures heterogeneous observations matrix.}

\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{output.file}{File or path and file to output.  If indicated output directory (if path indicated) does not exist, a warning is issued and program execution halts.  Default behavior is to write output to LRCDE_power.analysis.csv in the current working directory.}

\item{FEEDBACK}{Boolean indicating whether to output progess indication to console.  Default is TRUE.}

\item{medCntr}{Boolean indicating whether to mean center differential expression estimates.}

\item{stdz}{Boolean indicatin whether to scale differential expression estimates with their pooled adjusted standard deviation}

\item{nonNeg}{Boolean indicating whether to force cell type-specific estimates to be non-negative (TRUE) or not (FALSE).}

\item{method}{Only "dual" is implemented in this version. This should be one of "single", "dual" (csSAM method), or "ridge".  Default is "dual".  Specifies which type of regression deconvolution to perform.}

\item{direction}{Should be one of "two.sided", "up", or "down".  Which direction to test for cell type-specific expression changes.}
}
\value{
List containing data.frame (total.frame) of analysis results and a list of parameter values supplied to lrcde function (arg.used).
}
\description{
Call this function to run entire functionality of LRCDE.
}
\details{
The lrcde function is meant to be called directly by user.  It is the entry point for the cell type-specific differential expression power analysis.
This is mainly a wrapper script for the deconvolution step and subsequent call to power analysis function (do.decon.power).
Default behavior is to write output to LRCDE_power.analysis.csv in the current working directory.
}
\examples{
# Load lrcde library:
library( "lrcde" )

# User chosen working directory - Where your output .csv file will be found:
setwd(getwd()); getwd()

# Load data to work with:
# Select an ExpressionSet object from the CellMix package to work with:
library(CellMix)
library(GEOquery)

# Data sets in CellMix to select from:
gedDataInfo()
mix <- ExpressionMix("GSE19830")
p.sub = subset(pData(mix), select=c( Liver, Brain, Lung,  characteristics_ch1))
head( p.sub )
unique( p.sub$characteristics_ch1 )
het <- t(exprs(mix))            # Heterogeneous expression matrix
cell.props <- t(coef(mix))      # cell proportions matrix
dim(het)
dim(cell.props)
apply(cell.props, 2, sd)
# Cell proportion standard deviation > 0.06 across samples should produce reasonable differential detection power.
# Brain     Liver      Lung 
# 0.2613150 0.2859059 0.2837341
apply(cell.props, 1, sum)            # These sum perfectly to 1.  NOT suitable for single-step deconvolution.

# Create (artificially) two groups
p.sub.1 = p.sub[ (p.sub$Brain + p.sub$Lung) >  ( p.sub$Liver )     ,]
dim(p.sub.1)
g.1.names = rownames(p.sub.1)
het.1 = het[ rownames(het) \%in\% g.1.names, ]
dim(het.1)
het.2 = het[ !(rownames(het) \%in\% g.1.names), ]
dim(het.2)
cells.1 = cell.props[ rownames(cell.props)\%in\%g.1.names,  ]
cells.2 = cell.props[ !(rownames(cell.props)\%in\%g.1.names),  ]
all.equal(rownames(het.1), rownames(cells.1))
all.equal(rownames(het.2), rownames(cells.2))
het2use = rbind(het.1, het.2)
cell.props = rbind(cells.1, cells.2)
groups = c(rep(1, dim(het.1)[1]), rep(2, dim(het.2)[1]) )
table(groups)

# Now apply the lrcde function in order to deconvolve cell type-specific expression
# and perform power analysis on differential expression detection:

power.analysis.df = lrcde( het2use, cell.props, groups
                           , output.file = "LRCDE_power_analysis_results"
                           , FEEDBACK = TRUE, medCntr = FALSE, stdz = FALSE, nonNeg = TRUE
                           , method = "dual", direction = "two.sided")



}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

