% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrcde.R
\name{lrcde}
\alias{lrcde}
\title{lrcde}
\usage{
lrcde(het.sub, cell.props, groups, output.file = "LRCDE_power_analysis",
  FEEDBACK = TRUE, medCntr = FALSE, stdz = FALSE, nonNeg = TRUE,
  method = "dual", direction = "two.sided")
}
\arguments{
\item{het.sub}{Should be samples by genomic site (rows by columns).  The samples by genomic measures heterogeneous observations matrix.}

\item{cell.props}{Should be samples by cell types (rows by columns).  The relative cell proportions per sample.}

\item{groups}{A vector of 1's and 2's indicating group membership per sample.  Should align with samples in het.sub and cell.props.}

\item{output.file}{File or path and file to output.  If indicated output directory (if path indicated) does not exist, a warning is issued and program execution halts.  Default behavior is to write output to LRCDE_power.analysis.csv in the current working directory.}

\item{FEEDBACK}{Boolean indicating whether to output progess indication to console.  Default is TRUE.}

\item{medCntr}{Boolean indicating whether to mean center differential expression estimates.}

\item{stdz}{Boolean indicatin whether to scale differential expression estimates with their pooled adjusted standard deviation}

\item{nonNeg}{Boolean indicating whether to force cell type-specific estimates to be non-negative (TRUE) or not (FALSE).}

\item{method}{Only "dual" is implemented in this version. This should be one of "single", "dual" (csSAM method), or "ridge".  Default is "dual".  Specifies which type of regression deconvolution to perform.}

\item{direction}{Should be one of "two.sided", "up", or "down".  Which direction to test for cell type-specific expression changes.}
}
\value{
List containing data.frame (total.frame) of analysis results and a list of parameter values supplied to lrcde function (arg.used).
}
\description{
Call this function to run entire functionality of LRCDE.
}
\details{
The lrcde function is meant to be called directly by user.  It is the entry point for the cell type-specific differential expression power analysis.
This is mainly a wrapper script for the deconvolution step and subsequent call to power analysis function (do.decon.power).
Default behavior is to write output to LRCDE_power.analysis.csv in the current working directory.
}
\examples{
lrcde( het.sub, cell.props, groups, output.file="name.of.csv.file.to.create", method="dual", direction="two.sided" )
}
\author{
Edmund R Glass, \email{Edmund.Glass@gmail.com}, Mikhail G Dozmorov, \email{Mikhail.Dozmorov@vcuhealth.org}
}
\references{
\url{https://github.com/ERGlass/lrcde.dev}
}
\keyword{Deconvolution}
\keyword{analysis}
\keyword{cell}
\keyword{detection}
\keyword{differential}
\keyword{expression}
\keyword{power}
\keyword{type-specific}

